+++ bin/.gitignore
@@ -2,3 +2,4 @@
 /model/
 /controller/
 /view/
+/manager/
