+++ src/model/Ordine.java
@@ -2,12 +2,15 @@ package model;
 
 import javafx.beans.property.FloatProperty;
 import javafx.beans.property.SimpleFloatProperty;
+import javafx.beans.property.SimpleStringProperty;
+import javafx.beans.property.StringProperty;
 
 
 public class Ordine {
 	protected java.sql.Time orario;
 	protected java.sql.Date data;
 	private FloatProperty prezzoTotale;
+	protected StringProperty nomeCliente;
 	
 	
 	public Ordine() {
@@ -16,13 +19,15 @@ public class Ordine {
 		this.orario = new java.sql.Time(today.getTime());
 		this.data = new java.sql.Date(millis);
 		this.prezzoTotale = new SimpleFloatProperty();
+		this.nomeCliente = new SimpleStringProperty();
 	}
 	
 
-	public Ordine(java.sql.Time orario, java.sql.Date data) {
+	public Ordine(java.sql.Time orario, java.sql.Date data,Float prezzoTotale,StringProperty nomeCliente) {
 		this.setOrario(orario);
 		this.setDate(data);
-		
+		this.setPrezzoTotale(prezzoTotale);
+		this.setNomeCliente(nomeCliente);
 	}
 	
 	// metodi orario ordine
@@ -32,6 +37,14 @@ public class Ordine {
 
 	public void setOrario(java.sql.Time Time) {
 		this.orario = Time;
+	}
+	
+	public StringProperty getNomeCliente() {
+		return nomeCliente;
+	}
+
+	public void setNomeCliente(StringProperty nomeCliente) {
+		this.nomeCliente = nomeCliente;
 	}
 	
 	
@@ -50,11 +63,11 @@ public class Ordine {
 			return prezzoTotale;
 	}
 
-	public void setPrezzo(float prezzo) {
+	public void setPrezzoTotale(float prezzo) {
 			this.prezzoTotale.set(prezzo);
 	}
 		
-	public float getPrezzo() {
+	public float getPrezzoTotale() {
 			return this.prezzoTotale.get();
 	}
